requirements.txt
# Database connectivity
psycopg2-binary>=2.9.3

# Web framework
Flask>=2.0.1,<3.0.0
Werkzeug>=2.0.0,<3.0.0

# HTTP requests
requests>=2.25.0

# JWT Authentication
PyJWT>=2.4.0
bcrypt>=3.2.0

# Additional Flask dependencies (if needed)
click>=7.0
ItsDangerous>=2.0
Jinja2>=3.0
MarkupSafe>=2.0

# Development and debugging (optional)
# gunicorn>=20.1.0  # For production deployment
# python-dotenv>=0.19.0  # For environment variables
